test(dedupe): consolidate stateful_indicators tests (5 → 3 files)
Cluster d47c4f77bb07: merge similar_names test files for stateful_indicators

Files consolidated:
- test_stateful_indicators.py (NEW): bundle + zscore (144 lines)
- test_stateful_indicators_algorithm.py (NEW): rsi + sma_and_ema (187 lines)
- test_stateful_indicators_primitives.py: renamed from welford_and_window (118 lines)

Removed:
- test_stateful_indicators_bundle_and_performance.py
- test_stateful_indicators_rsi.py
- test_stateful_indicators_sma_and_ema.py
- test_stateful_indicators_zscore.py

All 33 tests pass.
