+++ tests/_triage/dedupe_candidates.yaml
@@ -2552,7 +2552,9 @@ clusters:
     expected_file_delta: -4
     priority: low
     rationale: Consolidate 5 test files into single parametrized file
-    status: pending
+    status: done
+    owner: claude
+    pr_url: https://github.com/Solders-Girdles/GPT-Trader/pull/319
     added: '2026-01-20'
   da0ea2376a4d:
     type: similar_names
