+++ tests/unit/gpt_trader/features/live_trade/strategies/test_stateful_indicators_primitives.py
